Removing usage of Query from URI.
I've made the executive decision that all of the parrs of the URI should
be stored as strings internally and extra functionality (like that
provided by Query) doesn't need to be 'baked in' to the URI object. Such
functionality can be utilized out-of-band and then reinserted as a
string when needed.