Testing URI-encoding example
I had to modify the example, because it wasn't put together at a point
when I had the full specs of the QueryDict in mind. I.e.

    query[key] += ' value'

does not tack 'value' to the end of the value returned by query[key]
because the re-assignment appends to the list instead, so you get:

    [ 'previous', 'previous value' ]

instead of:

    [ 'previous value' ]

Oops! But, alas, this is what testing is for (and a very good reason to
test your documenation examples too!).